PR: Delete the todo, move the fixtures a level up
- Only keep the required fixtured within each test package.
- Move the common schema, and runner function to one level above.
